Add debug configs
