--- out.csv
@@ -1,48 +1,0 @@
-name,vote,group,canton
-Knecht,Hansjörg,+,AG
-Kuprecht,Alex,+,SZ
-Maret,Marianne,+,VS
-Mazzone,Lisa,+,GE
-Michel,Matthias,+,ZG
-Minder,Thomas,+,SH
-Müller,Damian,=,LU
-Noser,Ruedi,=,ZH
-Reichmuth,Othmar,+,SZ
-Rieder,Beat,+,VS
-Salzmann,Werner,+,BE
-Schmid,Martin,+,GR
-Sommaruga,Carlo,+,GE
-Stark,Jakob,+,TG
-Stöckli,Hans,+,BE
-Thorens Goumaz,Adèle,+,VD
-Vara,Céline,+,NE
-Wicki,Hans,-,NW
-Würth,Benedikt,+,SG
-Zanetti,Roberto,+,SO
-Z'graggen,Heidi,+,UR
-Zopfi,Mathias,+,GL
-,,,
-Bauer,Philippe,+,NE
-Bischof,Pirmin,+,SO
-Burkart,Thierry,+,AG
-Caroni,Andrea,+,AR
-Chassot,Isabelle,+,FR
-Chiesa,Marco,+,TI
-Crevoisier Crelier,Mathilde,+,JU
-Dittli,Josef,=,UR
-Engler,Stefan,+,GR
-Ettlin,Erich,+,OW
-Fässler,Daniel,+,AI
-Français,Olivier,+,VD
-Friedli,Esther,+,SG
-Gapany,Johanna,+,FR
-Germann,Hannes,=,SH
-"Gmür-
-Schönenberger",Andrea,+,LU
-Graf,Maya,+,BL
-Häberli-Koller,Brigitte,P,TG
-Hefti,Thomas,+,GL
-Hegglin,Peter,+,ZG
-Herzog,Eva,+,BS
-Jositsch,Daniel,+,ZH
-Juillard,Charles,+,JU
