+++ pyproject.toml
@@ -20,4 +20,4 @@ requires = ["poetry-core"]
 build-backend = "poetry.core.masonry.api"
 
 [tool.poetry.scripts]
-extract-votes = "vote_extractor:main"
+extract-votes = "vote_extractor.__main__:main"
