--- recommendations.json
@@ -1,14 +1,0 @@
-[
-    {
-        "business": "25527",
-        "recommendation": "+"
-    },
-    {
-        "business": "25528",
-        "recommendation": "+"
-    },
-    {
-        "business": "25523",
-        "recommendation": "-"
-    }
-]
